# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AuthSys API",
    "version": "23.0.0",
    "description": "Private OAuth 2.0 and OpenID Connect provider. Authorization Code + PKCE is the supported browser flow."
  },
  "servers": [
    { "url": "https://testmono-user.duckdns.org", "description": "Current HTTPS deployment" },
    { "url": "http://localhost:8080", "description": "Local Docker fallback" }
  ],
...
  "paths": {
...
    "/api/auth/register": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Create a local account and browser session",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RegisterRequest" } } }
        },
        "responses": {
          "201": { "description": "Account and HttpOnly browser session created" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "409": { "description": "Username or email already exists" }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Sign in for the AuthSys web UI",
        "description": "Creates an HttpOnly session cookie. Tokens are not returned to browser JavaScript.",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LoginRequest" } } }
        },
        "responses": {
          "200": { "description": "Authenticated browser session", "headers": { "Set-Cookie": { "schema": { "type": "string" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
...
    "/api/sessions": {
      "get": {
        "tags": ["Sessions"],
        "summary": "List active sessions",
        "security": [{ "cookieAuth": [] }, { "bearerAuth": [] }],
        "responses": { "200": { "description": "Sanitized session list" } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": { "type": "apiKey", "in": "cookie", "name": "connect.sid" },
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" },
      "clientBasic": { "type": "http", "scheme": "basic", "description": "OAuth client_id and client_secret" }
    },
...
    "schemas": {
      "RegisterRequest": {
        "type": "object",
        "required": ["username", "email", "password", "consentEssential"],
        "properties": {
          "username": { "type": "string", "minLength": 3, "maxLength": 64 },
          "email": { "type": "string", "format": "email", "maxLength": 254 },
          "password": { "type": "string", "minLength": 8, "maxLength": 128 },
          "consentEssential": { "type": "boolean", "enum": [true] }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "format": "email", "maxLength": 254 },
          "password": { "type": "string", "maxLength": 128 },
          "remember": { "type": "boolean", "default": false }
        }
      },
...
    },
    "responses": {
      "BadRequest": { "description": "Invalid or unsupported request" },
      "Unauthorized": { "description": "Authentication or token validation failed" },
      "RateLimited": { "description": "Rate limit exceeded" }
    }
  }
}